requirements.txt
pytest
selene==2.0.0rc9
allure-pytest
jsonschema
requests


